+++ test/testarray.cpp
@@ -826,7 +826,7 @@ TEST(Greater) {
 		for(int i = 0; i < items; i++) {
 			a.Add(0);
 		}
-		size_t t = a.Query<GREATER>(0, 0, -1);
+		size_t t = a.Query<GREATER>(0, 0, (size_t)-1);
 		CHECK_EQUAL(-1, t);
 
 
@@ -836,7 +836,7 @@ TEST(Greater) {
 		}
 		for(int i = 0; i < items; i++) {
 			a.Set(i, 1);
-			size_t t = a.Query<GREATER>(0, 0, -1);
+			size_t t = a.Query<GREATER>(0, 0, (size_t)-1);
 			CHECK_EQUAL(i, t);
 			a.Set(i, 0);
 		}
@@ -847,7 +847,7 @@ TEST(Greater) {
 		}
 		for(int i = 0; i < items; i++) {
 			a.Set(i, 3);
-			size_t t = a.Query<GREATER>(2, 0, -1);
+			size_t t = a.Query<GREATER>(2, 0, (size_t)-1);
 			CHECK_EQUAL(i, t);
 			a.Set(i, 2);
 		}
@@ -858,7 +858,7 @@ TEST(Greater) {
 		}
 		for(int i = 0; i < items; i++) {
 			a.Set(i, 11);
-			size_t t = a.Query<GREATER>(10, 0, -1);
+			size_t t = a.Query<GREATER>(10, 0, (size_t)-1);
 			CHECK_EQUAL(i, t);
 			a.Set(i, 10);
 		}
@@ -869,7 +869,7 @@ TEST(Greater) {
 		}
 		for(int i = 0; i < items; i++) {
 			a.Set(i, 110);
-			size_t t = a.Query<GREATER>(100, 0, -1);
+			size_t t = a.Query<GREATER>(100, 0, (size_t)-1);
 			CHECK_EQUAL(i, t);
 			a.Set(i, 100);
 		}
@@ -879,7 +879,7 @@ TEST(Greater) {
 		}
 		for(int i = 0; i < items; i++) {
 			a.Set(i, 210);
-			size_t t = a.Query<GREATER>(200, 0, -1);
+			size_t t = a.Query<GREATER>(200, 0, (size_t)-1);
 			CHECK_EQUAL(i, t);
 			a.Set(i, 200);
 		}
@@ -890,7 +890,7 @@ TEST(Greater) {
 		}
 		for(int i = 0; i < items; i++) {
 			a.Set(i, 11000);
-			size_t t = a.Query<GREATER>(10000, 0, -1);
+			size_t t = a.Query<GREATER>(10000, 0, (size_t)-1);
 			CHECK_EQUAL(i, t);
 			a.Set(i, 10000);
 		}
@@ -901,7 +901,7 @@ TEST(Greater) {
 
 		for(int i = 0; i < items; i++) {
 			a.Set(i, 41000);
-			size_t t = a.Query<GREATER>(40000, 0, -1);
+			size_t t = a.Query<GREATER>(40000, 0, (size_t)-1);
 			CHECK_EQUAL(i, t);
 			a.Set(i, 40000);
 		}
@@ -912,7 +912,7 @@ TEST(Greater) {
 		}
 		for(int i = 0; i < items; i++) {
 			a.Set(i, 1100000);
-			size_t t = a.Query<GREATER>(1000000, 0, -1);
+			size_t t = a.Query<GREATER>(1000000, 0, (size_t)-1);
 			CHECK_EQUAL(i, t);
 			a.Set(i, 1000000);
 		}
@@ -947,7 +947,7 @@ TEST(Less) {
 		for(int i = 0; i < items; i++) {
 			a.Add(0);
 		}
-		size_t t = a.Query<LESS>(0, 0, -1);
+		size_t t = a.Query<LESS>(0, 0, (size_t)-1);
 		CHECK_EQUAL(-1, t);
 
 
@@ -957,7 +957,7 @@ TEST(Less) {
 		}
 		for(int i = 0; i < items; i++) {
 			a.Set(i, 0);
-			size_t t = a.Query<LESS>(1, 0, -1);
+			size_t t = a.Query<LESS>(1, 0, (size_t)-1);
 			CHECK_EQUAL(i, t);
 			a.Set(i, 1);
 		}
@@ -968,7 +968,7 @@ TEST(Less) {
 		}
 		for(int i = 0; i < items; i++) {
 			a.Set(i, 2);
-			size_t t = a.Query<LESS>(3, 0, -1);
+			size_t t = a.Query<LESS>(3, 0, (size_t)-1);
 			CHECK_EQUAL(i, t);
 			a.Set(i, 3);
 		}
@@ -979,7 +979,7 @@ TEST(Less) {
 		}
 		for(int i = 0; i < items; i++) {
 			a.Set(i, 10);
-			size_t t = a.Query<LESS>(11, 0, -1);
+			size_t t = a.Query<LESS>(11, 0, (size_t)-1);
 			CHECK_EQUAL(i, t);
 			a.Set(i, 11);
 		}
@@ -990,7 +990,7 @@ TEST(Less) {
 		}
 		for(int i = 0; i < items; i++) {
 			a.Set(i, 100);
-			size_t t = a.Query<LESS>(110, 0, -1);
+			size_t t = a.Query<LESS>(110, 0, (size_t)-1);
 			CHECK_EQUAL(i, t);
 			a.Set(i, 110);
 		}
@@ -1000,7 +1000,7 @@ TEST(Less) {
 		}
 		for(int i = 0; i < items; i++) {
 			a.Set(i, 200);
-			size_t t = a.Query<LESS>(210, 0, -1);
+			size_t t = a.Query<LESS>(210, 0, (size_t)-1);
 			CHECK_EQUAL(i, t);
 			a.Set(i, 210);
 		}
@@ -1011,7 +1011,7 @@ TEST(Less) {
 		}
 		for(int i = 0; i < items; i++) {
 			a.Set(i, 10000);
-			size_t t = a.Query<LESS>(11000, 0, -1);
+			size_t t = a.Query<LESS>(11000, 0, (size_t)-1);
 			CHECK_EQUAL(i, t);
 			a.Set(i, 11000);
 		}
@@ -1022,7 +1022,7 @@ TEST(Less) {
 
 		for(int i = 0; i < items; i++) {
 			a.Set(i, 40000);
-			size_t t = a.Query<LESS>(41000, 0, -1);
+			size_t t = a.Query<LESS>(41000, 0, (size_t)-1);
 			CHECK_EQUAL(i, t);
 			a.Set(i, 41000);
 		}
@@ -1033,7 +1033,7 @@ TEST(Less) {
 		}
 		for(int i = 0; i < items; i++) {
 			a.Set(i, 1000000);
-			size_t t = a.Query<LESS>(1100000, 0, -1);
+			size_t t = a.Query<LESS>(1100000, 0, (size_t)-1);
 			CHECK_EQUAL(i, t);
 			a.Set(i, 1100000);
 		}
@@ -1044,7 +1044,7 @@ TEST(Less) {
 		}
 		for(int i = 0; i < items; i++) {
 			a.Set(i, 1000ULL*1000ULL*1000ULL*1000ULL - 1ULL);
-			size_t t = a.Query<LESS>(1000ULL*1000ULL*1000ULL*1000ULL, 0, -1);
+			size_t t = a.Query<LESS>(1000ULL*1000ULL*1000ULL*1000ULL, 0, (size_t)-1);
 			CHECK_EQUAL(i, t);
 			a.Set(i, 1000ULL*1000ULL*1000ULL*1000ULL);
 		}
